@@ -1,6 +1,6 @@
 [tool.ruff]
 line-length = 120
-target-version = "py311"
+target-version = "py313"
 fix = true
 show-fixes = true
 [tool.ruff.lint]
