@@ -67,7 +67,7 @@ class Downloader(object):
         """Extract download link from web page."""
         raise NotImplementedError(implement_method)
 
-    def specific_version(self: Self, app: APP, version: str) -> tuple[str, str]:
+    def specific_version(self: Self, app: "APP", version: str) -> tuple[str, str]:
         """Function to download the specified version of app..
 
         :param app: Name of the application
