@@ -24,6 +24,9 @@ class Parser(object):
     OUTPUT_ARG = "-o"
     KEYSTORE_ARG = "--keystore"
     OPTIONS_ARG = "-O"
+    ENABLE_ARG = "-e"
+    DISABLE_ARG = "-d"
+    EXCLUSIVE_ARG = "--exclusive"
 
     def __init__(self: Self, patcher: Patches, config: RevancedConfig) -> None:
         self._PATCHES: list[str] = []
@@ -71,7 +74,7 @@ class Parser(object):
             for opt in options:
                 pair = self.format_option(opt)
                 self._PATCHES[:0] = [self.OPTIONS_ARG, pair]
-        self._PATCHES[:0] = ["-e", name]
+        self._PATCHES[:0] = [self.ENABLE_ARG, name]
 
     def exclude(self: Self, name: str) -> None:
         """The `exclude` function adds a given patch to the list of excluded patches.
@@ -81,7 +84,7 @@ class Parser(object):
         name : str
             The `name` parameter is a string that represents the name of the patch to be excluded.
         """
-        self._PATCHES.extend(["-d", name])
+        self._PATCHES.extend([self.DISABLE_ARG, name])
         self._EXCLUDED.append(name)
 
     def get_excluded_patches(self: Self) -> list[str]:
@@ -119,22 +122,27 @@ class Parser(object):
             name = name.lower().replace(" ", "-")
             indices = [i for i in range(len(self._PATCHES)) if self._PATCHES[i] == name]
             for patch_index in indices:
-                if self._PATCHES[patch_index - 1] == "-e":
-                    self._PATCHES[patch_index - 1] = "-d"
+                if self._PATCHES[patch_index - 1] == self.ENABLE_ARG:
+                    self._PATCHES[patch_index - 1] = self.DISABLE_ARG
                 else:
-                    self._PATCHES[patch_index - 1] = "-e"
+                    self._PATCHES[patch_index - 1] = self.ENABLE_ARG
         except ValueError:
             return False
         else:
             return True
 
-    def exclude_all_patches(self: Self) -> None:
-        """The function `exclude_all_patches` exclude all the patches."""
-        for idx, item in enumerate(self._PATCHES):
-            if idx == 0:
-                continue
-            if item == "-e":
-                self._PATCHES[idx] = "-d"
+    def enable_exclusive_mode(self: Self) -> None:
+        """Enable exclusive mode - only explicitly enabled patches will run, all others disabled by default."""
+        logger.info("Enabling exclusive mode for fast testing - only keeping one patch enabled.")
+        # Clear all patches and keep only the first one enabled
+        if self._PATCHES:
+            # Find the first enable argument and its patch name
+            for idx in range(0, len(self._PATCHES), 2):
+                if idx < len(self._PATCHES) and self._PATCHES[idx] == self.ENABLE_ARG and idx + 1 < len(self._PATCHES):
+                    first_patch = self._PATCHES[idx + 1]
+                    # Clear all patches and set only the first one
+                    self._PATCHES = [self.ENABLE_ARG, first_patch]
+                    break
 
     def fetch_patch_options(self: Self, name: str, options_list: list[dict[str, Any]]) -> dict[str, Any]:
         """The function `fetch_patch_options` finds patch options for the patch.
@@ -151,6 +159,201 @@ class Parser(object):
             {},
         )
 
+    def _load_options_from_file(self: Self, file_name: str) -> list[dict[str, Any]]:
+        """Load options from a single file.
+
+        Parameters
+        ----------
+        file_name : str
+            The options file name
+
+        Returns
+        -------
+        list[dict[str, Any]]
+            List of patch options from the file
+        """
+        try:
+            with self.config.temp_folder.joinpath(file_name).open() as file:
+                options: list[dict[str, Any]] = json.load(file)
+                return options
+        except FileNotFoundError as e:
+            logger.warning(str(e))
+            return []
+
+    def _merge_options(
+        self: Self,
+        global_options: list[dict[str, Any]],
+        app_options: list[dict[str, Any]],
+    ) -> list[dict[str, Any]]:
+        """Merge global and app-specific options.
+
+        App-specific options override global options for the same patch name.
+
+        Parameters
+        ----------
+        global_options : list[dict[str, Any]]
+            Options from the global options file
+        app_options : list[dict[str, Any]]
+            Options from the app-specific options file
+
+        Returns
+        -------
+        list[dict[str, Any]]
+            Merged options list
+        """
+        # Create a dict keyed by patchName for easy lookup and merging
+        merged: dict[str, dict[str, Any]] = {}
+
+        # Add global options first
+        for opt in global_options:
+            patch_name = opt.get("patchName")
+            if patch_name:
+                merged[patch_name] = opt
+
+        # Override/add with app-specific options
+        for opt in app_options:
+            patch_name = opt.get("patchName")
+            if patch_name:
+                merged[patch_name] = opt
+
+        return list(merged.values())
+
+    def _load_patch_options(self: Self, app: APP) -> list[dict[str, Any]]:
+        """Load patch options from file.
+
+        Loads global options first, then merges app-specific options on top.
+        App-specific options override global options for the same patch name.
+
+        Parameters
+        ----------
+        app : APP
+            The app instance
+
+        Returns
+        -------
+        list[dict[str, Any]]
+            List of patch options
+        """
+        # Load global options first
+        global_options = self._load_options_from_file(self.config.global_options_file)
+
+        # If app uses a different options file, load and merge it
+        if app.options_file != self.config.global_options_file:
+            logger.info(f"Loading app-specific options from {app.options_file} and merging with global options")
+            app_options = self._load_options_from_file(app.options_file)
+            return self._merge_options(global_options, app_options)
+
+        return global_options if global_options else [{}]
+
+    def _normalize_patch_name(self: Self, patch_name: str, *, space_formatted: bool) -> str:
+        """Normalize patch name based on formatting preference.
+
+        Parameters
+        ----------
+        patch_name : str
+            The original patch name
+        space_formatted : bool
+            Whether to use space formatting
+
+        Returns
+        -------
+        str
+            Normalized patch name
+        """
+        return patch_name.lower().replace(" ", "-") if space_formatted else patch_name
+
+    def _should_include_regular_patch(self: Self, patch_name: str, normalized_name: str, app: APP) -> bool:
+        """Determine if a regular patch should be included.
+
+        Parameters
+        ----------
+        patch_name : str
+            The original patch name
+        normalized_name : str
+            The normalized patch name
+        app : APP
+            The app instance
+
+        Returns
+        -------
+        bool
+            True if patch should be included
+        """
+        exclude_list = app.exclude_request
+        check_name = normalized_name if app.space_formatted else patch_name
+        return check_name not in exclude_list
+
+    def _should_include_universal_patch(self: Self, patch_name: str, normalized_name: str, app: APP) -> bool:
+        """Determine if a universal patch should be included.
+
+        Parameters
+        ----------
+        patch_name : str
+            The original patch name
+        normalized_name : str
+            The normalized patch name
+        app : APP
+            The app instance
+
+        Returns
+        -------
+        bool
+            True if patch should be included
+        """
+        include_list = app.include_request
+        check_name = normalized_name if app.space_formatted else patch_name
+        return check_name in include_list
+
+    def _process_regular_patches(
+        self: Self,
+        patches: list[dict[str, str]],
+        app: APP,
+        options_list: list[dict[str, Any]],
+    ) -> None:
+        """Process regular patches for include/exclude.
+
+        Parameters
+        ----------
+        patches : list[dict[str, str]]
+            List of regular patches
+        app : APP
+            The app instance
+        options_list : list[dict[str, Any]]
+            List of patch options
+        """
+        for patch in patches:
+            patch_name = patch["name"]
+            normalized_name = self._normalize_patch_name(patch_name, space_formatted=app.space_formatted)
+
+            if self._should_include_regular_patch(patch_name, normalized_name, app):
+                self.include(patch_name, options_list)
+            else:
+                self.exclude(patch_name)
+
+    def _process_universal_patches(
+        self: Self,
+        universal_patches: list[dict[str, str]],
+        app: APP,
+        options_list: list[dict[str, Any]],
+    ) -> None:
+        """Process universal patches for include.
+
+        Parameters
+        ----------
+        universal_patches : list[dict[str, str]]
+            List of universal patches
+        app : APP
+            The app instance
+        options_list : list[dict[str, Any]]
+            List of patch options
+        """
+        for patch in universal_patches:
+            patch_name = patch["name"]
+            normalized_name = self._normalize_patch_name(patch_name, space_formatted=app.space_formatted)
+
+            if self._should_include_universal_patch(patch_name, normalized_name, app):
+                self.include(patch_name, options_list)
+
     def include_exclude_patch(
         self: Self,
         app: APP,
@@ -158,46 +361,59 @@ class Parser(object):
         patches_dict: dict[str, list[dict[str, str]]],
     ) -> None:
         """The function `include_exclude_patch` includes and excludes patches for a given app."""
-        options_list: list[dict[str, Any]] = [{}]
-        try:
-            with self.config.temp_folder.joinpath(app.options_file).open() as file:
-                options_list = json.load(file)
-        # Not excepting on JSONDecodeError as it should error out if the file is not a valid JSON
-        except FileNotFoundError as e:
-            logger.warning(str(e))
-            logger.debug("Setting options to empty list.")
+        options_list = self._load_patch_options(app)
 
-        if app.space_formatted:
-            for patch in patches:
-                normalized_patch = patch["name"].lower().replace(" ", "-")
-                (
-                    self.include(patch["name"], options_list)
-                    if normalized_patch not in app.exclude_request
-                    else self.exclude(
-                        patch["name"],
-                    )
-                )
-            for patch in patches_dict["universal_patch"]:
-                normalized_patch = patch["name"].lower().replace(" ", "-")
-                (
-                    self.include(
-                        patch["name"],
-                        options_list,
-                    )
-                    if normalized_patch in app.include_request
-                    else ()
-                )
+        self._process_regular_patches(patches, app, options_list)
+        self._process_universal_patches(patches_dict["universal_patch"], app, options_list)
+
+    def _build_base_args(self: Self, app: APP) -> list[str]:
+        """Build base arguments for ReVanced CLI."""
+        return [
+            self.CLI_JAR,
+            app.resource["cli"]["file_name"],
+            self.NEW_APK_ARG,
+            app.download_file_name,
+        ]
+
+    def _add_patch_bundles(self: Self, args: list[str], app: APP) -> None:
+        """Add patch bundle arguments to the command."""
+        if hasattr(app, "patch_bundles") and app.patch_bundles:
+            # Use multiple -p arguments for multiple bundles
+            for bundle in app.patch_bundles:
+                args.extend([self.PATCHES_ARG, bundle["file_name"]])
         else:
-            for patch in patches:
-                (
-                    self.include(patch["name"], options_list)
-                    if patch["name"] not in app.exclude_request
-                    else self.exclude(
-                        patch["name"],
-                    )
-                )
-            for patch in patches_dict["universal_patch"]:
-                self.include(patch["name"], options_list) if patch["name"] in app.include_request else ()
+            # Fallback to single bundle for backward compatibility
+            args.extend([self.PATCHES_ARG, app.resource["patches"]["file_name"]])
+
+    def _add_output_and_keystore_args(self: Self, args: list[str], app: APP) -> None:
+        """Add output file and keystore arguments."""
+        args.extend(
+            [
+                self.OUTPUT_ARG,
+                app.get_output_file_name(),
+                self.KEYSTORE_ARG,
+                app.keystore_name,
+                "--force",
+            ],
+        )
+
+    def _add_keystore_flags(self: Self, args: list[str], app: APP) -> None:
+        """Add keystore-specific flags if needed."""
+        if app.old_key:
+            # https://github.com/ReVanced/revanced-cli/issues/272#issuecomment-1740587534
+            old_key_flags = [
+                "--keystore-entry-alias=alias",
+                "--keystore-entry-password=ReVanced",
+                "--keystore-password=ReVanced",
+            ]
+            args.extend(old_key_flags)
+
+    def _add_architecture_args(self: Self, args: list[str], app: APP) -> None:
+        """Add architecture-specific arguments."""
+        if app.app_name in self.config.rip_libs_apps:
+            excluded = set(possible_archs) - set(app.archs_to_build)
+            for arch in excluded:
+                args.extend(("--rip-lib", arch))
 
     # noinspection IncorrectFormatting
     def patch_app(
@@ -212,39 +428,23 @@ class Parser(object):
             The `app` parameter is an instance of the `APP` class. It represents an application that needs
         to be patched.
         """
-        apk_arg = self.NEW_APK_ARG
-        exp = "--force"
-        args = [
-            self.CLI_JAR,
-            app.resource["cli"]["file_name"],
-            apk_arg,
-            app.download_file_name,
-            self.PATCHES_ARG,
-            app.resource["patches"]["file_name"],
-            self.OUTPUT_ARG,
-            app.get_output_file_name(),
-            self.KEYSTORE_ARG,
-            app.keystore_name,
-            exp,
-        ]
-        args[1::2] = map(self.config.temp_folder.joinpath, args[1::2])
-        if app.old_key:
-            # https://github.com/ReVanced/revanced-cli/issues/272#issuecomment-1740587534
-            old_key_flags = [
-                "--keystore-entry-alias=alias",
-                "--keystore-entry-password=ReVanced",
-                "--keystore-password=ReVanced",
-            ]
-            args.extend(old_key_flags)
+        args = self._build_base_args(app)
+        self._add_patch_bundles(args, app)
+        self._add_output_and_keystore_args(args, app)
+
+        # Convert paths to absolute paths
+        args[1::2] = [str(self.config.temp_folder.joinpath(arg)) for arg in args[1::2]]
+
+        self._add_keystore_flags(args, app)
+
         if self.config.ci_test:
-            self.exclude_all_patches()
+            self.enable_exclusive_mode()
         if self._PATCHES:
             args.extend(self._PATCHES)
-        if app.app_name in self.config.rip_libs_apps:
-            excluded = set(possible_archs) - set(app.archs_to_build)
-            for arch in excluded:
-                args.extend(("--rip-lib", arch))
+
+        self._add_architecture_args(args, app)
         args.extend(("--purge",))
+
         start = perf_counter()
         logger.debug(f"Sending request to revanced cli for building with args java {args}")
         process = Popen(["java", *args], stdout=PIPE)
